set SendMoneyApp to AndroidManifest and refactor StringUtil and create MoneyUtil classes

+++ app/src/main/java/br/com/example/sendmoney/model/entity/Transfer.kt
@@ -1,3 +1,5 @@
 package br.com.example.sendmoney.model.entity
 
-data class Transfer(var id: Int?, var name: String?, var phone: String?, var value: Double?)
+import java.math.BigDecimal
+
+data class Transfer(var id: Int?, var name: String?, var phone: String?, var value: BigDecimal?)
